chore: Update file provider to include file triggers in collection definition

+++ packages/server/src/services/file/service.js
@@ -1,6 +1,7 @@
 const fs = require("file-system");
 const pathModule = require("path");
 const DataProvider = require("./../data_provider/service");
+const { config } = require("./../../config");
 
 class FileService {
   constructor() {
@@ -14,6 +15,7 @@ class FileService {
   /**
    *
    * @param {string} fileType
+   * @param {string} tag
    *
    * @returns storedFile
    * @returns storedFile.fileName
@@ -22,21 +24,32 @@ class FileService {
    * @returns storedFile.fileFormat
    */
   createStoredDetail(fileType, tag) {
-    let time = new Date().getTime();
-    let fileFormat = fileType.split("/")[1];
-    let fileName = `${time}.${fileFormat}`;
-    let fullPath = pathModule.join(this.directory, fileFormat, tag, fileName);
+    const time = new Date().getTime();
+    const fileFormat = fileType.split("/")[1];
+    const fileName = `${time}.${fileFormat}`;
+    const fullPath = pathModule.join(this.directory, fileFormat, tag, fileName);
 
     return { fileName, fullPath, fileFormat };
   }
 
   /**
+   * Stores a file, removes the given temporary file, and submits file details into the database.
    *
-   * @param args
-   * @param {file} args.file file object
-   * @param {string} args.ownerId file object
+   * @param {Object} options - The options for storing the file.
+   * @param {Object} options.file - The file to be stored.
+   * @param {string} options.file.path - The path of the file.
+   * @param {string} options.file.type - The type of the file.
+   * @param {string} options.file.name - The original name of the file.
+   * @param {number} options.file.size - The size of the file.
+   * @param {string} options.ownerId - The ID of the owner of the file.
+   * @param {string} options.tag - The tag associated with the file.
+   * @param {boolean} [options.removeFileAfterStore=true] - Whether to remove the file after storing it.
+   *
+   * @returns {Promise} A promise that resolves with the document of the stored file.
+   *
+   * @throws {string} If the upload directory has not been set.
    */
-  storeFile({ file, ownerId, tag }) {
+  storeFile({ file, ownerId, tag, removeFileAfterStore = true }) {
     if (!this.directory) throw "upload directory has not been set.";
 
     let storedFile;
@@ -53,7 +66,7 @@ class FileService {
             else done();
 
             // remove temp file
-            fs.fs.unlinkSync(file.path);
+            if (removeFileAfterStore) fs.fs.unlinkSync(file.path);
           },
         });
       })
@@ -85,6 +98,12 @@ class FileService {
     );
   }
 
+  /**
+   * Removes a file from the disk.
+   *
+   * @param {string} path - The path of the file to be removed.
+   * @returns {Promise} A promise that resolves if the file is successfully removed, and rejects if an error occurs.
+   */
   removeFromDisc(path) {
     return new Promise((done, reject) => {
       fs.fs.unlink(path, (err) => {
@@ -94,6 +113,13 @@ class FileService {
     });
   }
 
+  /**
+   * Removes a file from the database and the disk.
+   *
+   * @param {string} fileId - The ID of the file to be removed.
+   * @returns {Promise} A promise that resolves if the file is successfully removed, and rejects if an error occurs.
+   * @throws Will throw an error if upload directory has not been set.
+   */
   removeFile(fileId) {
     if (!this.directory) throw "upload directory has not been set.";
 
@@ -128,6 +154,41 @@ class FileService {
         .then(done)
         .catch(reject);
     });
+  }
+
+  /**
+   * Retrieves a file from the database.
+   *
+   * @param {string} fileId - The ID of the file to be retrieved.
+   * @returns {Promise} A promise that resolves with the file document, or rejects if an error occurs.
+   */
+  getFile(fileId) {
+    const CollectionModel = DataProvider.getCollection("cms", "file");
+
+    return CollectionModel.findOne({ _id: fileId }).exec();
+  }
+
+  /**
+   * Retrieves the link of a file.
+   *
+   * @param {string} fileId - The ID of the file to get the link for.
+   * @returns {Promise} A promise that resolves with the file link, or rejects if an error occurs.
+   */
+  async getFileLink(fileId) {
+    const fileDoc = await this.getFile(fileId);
+
+    const link =
+      config.staticPath.rootPath +
+      `/${fileDoc.format}/${fileDoc.tag}/` +
+      fileDoc.fileName;
+
+    return link;
+  }
+
+  async getFilePath(fileId) {
+    const { fileName, format, tag } = await this.getFile(fileId);
+    const fullPath = pathModule.join(this.directory, format, tag, fileName);
+    return fullPath;
   }
 }
